moved files around a bit
now table scanner is a ui component,  it will only serve as one part of the ScanContacts page.
