+++ UIComponents/TableScanner/ScannerDrawer.js
